pyproject.toml
[project]
name = "labelingo"
version = "0.1.0"
description = "Annotate UI screenshots with translations"
authors = [
    { name = "Oliver Steele", email = "steele@osteele.com" }
]
dependencies = [
    "click>=8.0.0",
    "Pillow>=10.0.0",
    "anthropic>=0.7.0",
    "python-dotenv>=1.0.0",
    "easyocr>=1.7.0",
    "openai>=1.0.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = { file = "LICENSE" }

[project.optional-dependencies]
ocr = [
    "pytesseract>=0.3.0",
    "paddlepaddle>=2.5.0",
    "paddleocr>=2.7.0"
]
cairo = [
    "cairosvg>=2.7.0",
    "pycairo>=1.25.0",
    "cairocffi>=1.6.0",
]
dev = [
    "pre-commit>=3.6.0",
    "pyright>=1.1.350",
    "ruff>=0.3.3",
]

[project.scripts]
labelingo = "labelingo.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
target-version = "py38"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.hatch.build.targets.wheel]
packages = ["labelingo"]

[tool.uv]
dev-dependencies = [
    "pyright>=1.1.389",
    "ruff>=0.7.4",
]

[tool.pyright]
include = ["labelingo"]
stubPath = "labelingo/stubs"
typeCheckingMode = "basic"

[tool.hatch.envs.default]
dependencies = [
    "labelingo[dev]"
]

[tool.hatch.envs.default.scripts]
install = "uv pip install -e ."
install-dev = "uv pip install -e .[dev]"
install-cairo = "uv pip install -e .[cairo]"
install-all = "uv pip install -e .[dev,cairo,ocr]"

fix = [
    "ruff format .",
    "ruff check --fix ."
]
fmt = "ruff format ."
lint = [
    "ruff check .",
    "pyright"
]
typecheck = "pyright"

[tool.hatch.envs.cairo]
dependencies = [
    "labelingo[cairo]"
]

[tool.hatch.envs.ocr]
dependencies = [
    "labelingo[ocr]"
]

[tool.hatch.envs.full]
dependencies = [
    "labelingo[dev,cairo,ocr]"
]
